+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased][]
 
 -   Add support for managing side effects
+-   Make subviews, delegated events and side effects storage "private"
+    properties
 
 ## [6.0.1][] - 2019-04-04